Centralized utils to common_utils.h
