+++ include/awkward/Identities.h
@@ -8,7 +8,7 @@
 #include <map>
 #include <memory>
 
-#include "awkward/cpu-kernels/util.h"
+#include "awkward/common_utils.h"
 #include "awkward/Index.h"
 
 namespace awkward {
