+++ include/awkward/array/RawArray.h
@@ -12,7 +12,7 @@
 #include <stdexcept>
 #include <typeinfo>
 
-#include "awkward/cpu-kernels/util.h"
+#include "awkward/common_utils.h"
 #include "awkward/cpu-kernels/identities.h"
 #include "awkward/cpu-kernels/getitem.h"
 #include "awkward/cpu-kernels/operations.h"
