+++ include/awkward/builder/ArrayBuilder.h
@@ -3,7 +3,7 @@
 #ifndef AWKWARD_ARRAYBUILDER_H_
 #define AWKWARD_ARRAYBUILDER_H_
 
-#include "awkward/cpu-kernels/util.h"
+#include "awkward/common_utils.h"
 #include "awkward/Content.h"
 #include "awkward/type/Type.h"
 #include "awkward/builder/ArrayBuilderOptions.h"
