+++ include/awkward/builder/Builder.h
@@ -6,7 +6,7 @@
 #include <string>
 #include <vector>
 
-#include "awkward/cpu-kernels/util.h"
+#include "awkward/common_utils.h"
 #include "awkward/Content.h"
 #include "awkward/type/Type.h"
