+++ include/awkward/builder/Float64Builder.h
@@ -3,7 +3,7 @@
 #ifndef AWKWARD_FLOAT64BUILDER_H_
 #define AWKWARD_FLOAT64BUILDER_H_
 
-#include "awkward/cpu-kernels/util.h"
+#include "awkward/common_utils.h"
 #include "awkward/builder/ArrayBuilderOptions.h"
 #include "awkward/builder/GrowableBuffer.h"
 #include "awkward/builder/Builder.h"
