+++ include/awkward/cpu-kernels/reducers.h
@@ -3,7 +3,7 @@
 #ifndef AWKWARDCPU_REDUCERS_H_
 #define AWKWARDCPU_REDUCERS_H_
 
-#include "awkward/cpu-kernels/util.h"
+#include "awkward/common_utils.h"
 
 extern "C" {
   EXPORT_SYMBOL struct Error
