+++ include/awkward/kernel.h
@@ -1,7 +1,7 @@
 #ifndef AWKWARD_KERNEL_H
 #define AWKWARD_KERNEL_H
 
-#include "awkward/cpu-kernels/util.h"
+#include "awkward/common_utils.h"
 
 namespace kernel {
