+++ include/awkward/type/Type.h
@@ -6,7 +6,7 @@
 #include <memory>
 #include <vector>
 
-#include "awkward/cpu-kernels/util.h"
+#include "awkward/common_utils.h"
 #include "awkward/util.h"
 
 namespace awkward {
